ml_functions/requirements.txt
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-bigquery==3.14.1
google-cloud-core==2.4.1
google-cloud-scheduler==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
google-cloud-pubsub==2.16.0
